requirement.txt
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
mutagen==1.47.0
pycryptodomex==3.20.0
python-dotenv==1.0.1
pytube==15.0.0
requests==2.31.0
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.3.10
